HW: Change u8 pad specifier params to int where applicable

+++ Source/Core/Core/HW/GCPad.cpp
@@ -46,17 +46,17 @@ void LoadConfig()
   s_config.LoadConfig(true);
 }
 
-GCPadStatus GetStatus(u8 pad_num)
+GCPadStatus GetStatus(int pad_num)
 {
   return static_cast<GCPad*>(s_config.GetController(pad_num))->GetInput();
 }
 
-void Rumble(const u8 pad_num, const ControlState strength)
+void Rumble(const int pad_num, const ControlState strength)
 {
   static_cast<GCPad*>(s_config.GetController(pad_num))->SetOutput(strength);
 }
 
-bool GetMicButton(const u8 pad_num)
+bool GetMicButton(const int pad_num)
 {
   return static_cast<GCPad*>(s_config.GetController(pad_num))->GetMicButton();
 }
